% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visuals-plot_models_radar.R
\name{apply_color_scheme}
\alias{apply_color_scheme}
\title{Apply Color Scheme to ggplot}
\usage{
apply_color_scheme(scheme = "default", is_fill = TRUE)
}
\arguments{
\item{scheme}{Character name of color scheme}

\item{is_fill}{Logical, whether to apply fill scale (TRUE) or color scale (FALSE)}
}
\value{
ggplot2 scale object
}
\description{
Adds appropriate color/fill scales to a ggplot object.
}
\keyword{internal}
