% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariates-clustering.R
\name{assign_unknowns_to_clusters}
\alias{assign_unknowns_to_clusters}
\title{Assign Unknown Samples to OSSL Clusters}
\usage{
assign_unknowns_to_clusters(unknown_pca_scores, cluster_model, verbose = TRUE)
}
\arguments{
\item{unknown_pca_scores}{Tibble with unknown PCA scores (Dim.1, Dim.2, ...)}

\item{cluster_model}{List from cluster_ossl_samples() containing kmeans_model}

\item{verbose}{Logical. Print progress (default: TRUE)}
}
\value{
Integer vector of cluster assignments (length = nrow(unknown_pca_scores))
}
\description{
Matches unknown samples to nearest OSSL cluster centroids using Euclidean
distance in PCA space. Each unknown is assigned to the cluster whose centroid
it is closest to.
}
\keyword{internal}
