% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-backtransform.R
\name{back_transform_last_fit}
\alias{back_transform_last_fit}
\title{Back-transform Predictions in tune::last_fit Results}
\usage{
back_transform_last_fit(last_fit_results, transformation)
}
\arguments{
\item{last_fit_results}{Results object from tune::last_fit()}

\item{transformation}{Character string: "none", "log", or "sqrt"}
}
\value{
Modified last_fit results with back-transformed predictions
}
\description{
Applies back-transformation to predictions from tune::last_fit() results.
This is used in evaluation-core.R after model fitting.
}
\keyword{internal}
