% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-progress_reporting.R
\name{calculate_chunk_metrics}
\alias{calculate_chunk_metrics}
\title{Calculate Performance Metrics from Chunk Results}
\usage{
calculate_chunk_metrics(chunk_results, metrics = c("rrmse", "rsq"))
}
\arguments{
\item{chunk_results}{List of results from parallel chunk processing}

\item{metrics}{Character vector of metrics to calculate (default: c("rrmse", "rsq"))}
}
\value{
List containing best/avg/worst for each requested metric
}
\description{
Calculate Performance Metrics from Chunk Results
}
