% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-parallel_logging.R
\name{calculate_eta_with_ci}
\alias{calculate_eta_with_ci}
\title{Calculate ETA with Confidence Intervals}
\usage{
calculate_eta_with_ci(
  n_completed,
  n_remaining,
  elapsed_mins,
  model_times = NULL
)
}
\arguments{
\item{n_completed}{Number of completed models}

\item{n_remaining}{Number of remaining models}

\item{elapsed_mins}{Total elapsed time in minutes}

\item{model_times}{Vector of individual model completion times}
}
\value{
List with eta_median, eta_lower, eta_upper
}
\description{
Calculates estimated time to completion with confidence intervals
based on recent model completion times
}
