% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config-hpc.R
\name{calculate_latency_score}
\alias{calculate_latency_score}
\title{Calculate Latency Score}
\usage{
calculate_latency_score(candidates, avg_tuning_time_minutes)
}
\description{
Calculate Latency Score
}
\keyword{internal}
