% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config-nested_parallel.R
\name{calculate_throughput_score}
\alias{calculate_throughput_score}
\title{Calculate Throughput Score}
\usage{
calculate_throughput_score(
  candidates,
  n_models,
  memory_per_model_gb,
  system_memory_gb
)
}
\description{
Calculate Throughput Score
}
\keyword{internal}
