% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-resource_management.R
\name{check_resource_availability}
\alias{check_resource_availability}
\title{Check Resource Availability}
\usage{
check_resource_availability(resource_manager, n_models)
}
\arguments{
\item{resource_manager}{Resource manager object}

\item{n_models}{Number of models in the batch}
}
\description{
Verifies sufficient resources are available before starting a batch.
}
\keyword{internal}
