% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-resource_management.R
\name{check_zombie_processes}
\alias{check_zombie_processes}
\title{Check for Zombie Processes}
\usage{
check_zombie_processes(resource_manager)
}
\description{
Check for Zombie Processes
}
\keyword{internal}
