% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visuals-plot_component_value_interactions.R
\name{clean_component_names_advanced}
\alias{clean_component_names_advanced}
\title{Clean Component Names Advanced}
\usage{
clean_component_names_advanced(component_names)
}
\arguments{
\item{component_names}{Character vector of component names}
}
\value{
Character vector of cleaned names
}
\description{
Cleans component names for publication-quality display with line breaks.
}
\keyword{internal}
