% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-checkpoint_management.R
\name{cleanup_old_checkpoints}
\alias{cleanup_old_checkpoints}
\title{Clean Up Old Checkpoint Files}
\usage{
cleanup_old_checkpoints(checkpoint_dir, dry_run = TRUE)
}
\arguments{
\item{checkpoint_dir}{Character. Directory containing checkpoint files}

\item{dry_run}{Logical. If TRUE, only reports what would be deleted}
}
\value{
Number of files removed
}
\description{
Removes old individual checkpoint files if migrating from the old system
to the new batched system.
}
