% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-resource_management.R
\name{cleanup_resource_manager}
\alias{cleanup_resource_manager}
\title{Cleanup Resource Manager}
\usage{
cleanup_resource_manager(resource_manager)
}
\arguments{
\item{resource_manager}{Resource manager object}
}
\description{
Performs cleanup when resource manager is no longer needed.
}
\keyword{internal}
