% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{constants
Valid model types supported by the package}
\alias{constants
Valid model types supported by the package}
\alias{VALID_MODELS}
\title{Package Constants}
\format{
An object of class \code{character} of length 9.
}
\usage{
VALID_MODELS
}
\description{
Central location for package-wide constants used in validation and configuration.
}
\keyword{internal}
