% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{DEFAULT_GRID_SIZE}
\alias{DEFAULT_GRID_SIZE}
\title{Default tuning parameters}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
DEFAULT_GRID_SIZE
}
\description{
Default tuning parameters
}
\keyword{internal}
