% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-parallel_monitoring.R
\name{.read_progress_lines}
\alias{.read_progress_lines}
\title{(internal) Read and merge progress lines}
\usage{
.read_progress_lines(path)
}
\description{
(internal) Read and merge progress lines
}
\keyword{internal}
