% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-error_recovery.R
\name{execute_recovery_action}
\alias{execute_recovery_action}
\title{Execute Recovery Action}
\usage{
execute_recovery_action(action, resource_manager = NULL)
}
\arguments{
\item{action}{Character string action to execute}

\item{resource_manager}{Resource manager object}
}
\description{
Executes specific recovery actions based on strategy.
}
\keyword{internal}
