% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-progress_reporting.R
\name{format_global_best}
\alias{format_global_best}
\title{Format Global Best Model Information}
\usage{
format_global_best(all_results)
}
\arguments{
\item{all_results}{List of all results from parallel processing}
}
\value{
List containing best model information for display, or NULL if no successful models
}
\description{
Format Global Best Model Information
}
