% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline-configure.R
\name{generate_config_id}
\alias{generate_config_id}
\title{Generate a deterministic config ID}
\usage{
generate_config_id(model, preprocessing, transformation, feature_selection)
}
\arguments{
\item{model}{Character. Model short name.}

\item{preprocessing}{Character. Preprocessing method.}

\item{transformation}{Character. Response transformation.}

\item{feature_selection}{Character. Feature selection method.}
}
\value{
Character. Config ID in format \verb{\{model\}_\{preprocessing\}_\{transformation\}_\{feature_selection\}_\{6-char hash\}}.
}
\description{
Generate a deterministic config ID
}
\keyword{internal}
