% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visuals-plot_models_radar.R
\name{get_color_scheme}
\alias{get_color_scheme}
\title{Get Color Scheme for Plots}
\usage{
get_color_scheme(scheme = "default", n = 1)
}
\arguments{
\item{scheme}{Character name of color scheme}

\item{n}{Number of colors needed}
}
\value{
Character vector of hex colors
}
\description{
Returns color values based on the specified color scheme.
}
\keyword{internal}
