% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation-local-helpers.R
\name{get_completed_models}
\alias{get_completed_models}
\title{Get Completed Models}
\usage{
get_completed_models(checkpoint_dir)
}
\arguments{
\item{checkpoint_dir}{Directory containing checkpoints}
}
\value{
Character vector of completed model IDs
}
\description{
Retrieves list of completed models from checkpoint directory
}
\keyword{internal}
