% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visuals-plot_component_chord_diagram.R
\name{get_component_sector}
\alias{get_component_sector}
\title{Get Component Sector Assignment}
\usage{
get_component_sector(component_types)
}
\arguments{
\item{component_types}{Character vector of component types}
}
\value{
Character vector of sector names
}
\description{
Assigns components to logical sectors for chord diagram organization.
}
\keyword{internal}
