% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-resource_management.R
\name{get_current_memory_gb}
\alias{get_current_memory_gb}
\title{Get Current Memory Usage in GB}
\usage{
get_current_memory_gb()
}
\description{
Get Current Memory Usage in GB
}
\keyword{internal}
