% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor-nested_performance.R
\name{get_current_system_stats}
\alias{get_current_system_stats}
\title{Get Current System Statistics}
\usage{
get_current_system_stats()
}
\description{
Get Current System Statistics
}
\keyword{internal}
