% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-progress_reporting.R
\name{get_metrics_for_progress_level}
\alias{get_metrics_for_progress_level}
\title{Get Appropriate Metrics List Based on Progress Level}
\usage{
get_metrics_for_progress_level(progress_level = "standard")
}
\arguments{
\item{progress_level}{Character. One of "minimal", "standard", "detailed"}
}
\value{
Character vector of metrics to calculate and display
}
\description{
Get Appropriate Metrics List Based on Progress Level
}
