% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation-hpc-fixed.R
\name{get_optimal_workers}
\alias{get_optimal_workers}
\title{Get Optimal Workers}
\usage{
get_optimal_workers(context, level = "outer")
}
\description{
Determine optimal number of workers based on context.
}
\keyword{internal}
