% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_find_cache_path.R
\name{get_ossl_data_path}
\alias{get_ossl_data_path}
\title{Get Path to Cached OSSL Data}
\usage{
get_ossl_data_path(type = c("location", "lab", "mir"))
}
\arguments{
\item{type}{One of "location", "lab", or "mir".}
}
\value{
Path to the cached file or NULL if not present.
}
\description{
Returns the path to a specific cached file if it exists.
}
\keyword{internal}
