% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-resource_management.R
\name{get_peak_memory}
\alias{get_peak_memory}
\title{Get Peak Memory Usage}
\usage{
get_peak_memory(token)
}
\arguments{
\item{token}{Resource token}
}
\value{
Peak memory in GB
}
\description{
Returns the peak memory usage for a resource token.
}
\keyword{internal}
