% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visuals-plot_component_chord_diagram.R
\name{get_sector_order}
\alias{get_sector_order}
\title{Get Sector Display Order}
\usage{
get_sector_order(component_types)
}
\arguments{
\item{component_types}{Character vector of component types}
}
\value{
Numeric vector indicating sector order
}
\description{
Defines the order of sectors around the chord diagram circle.
}
\keyword{internal}
