% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-parallel_monitoring.R
\name{get_worker_statistics}
\alias{get_worker_statistics}
\title{Get Worker Statistics from Log}
\usage{
get_worker_statistics(log_file)
}
\arguments{
\item{log_file}{Path to log file}
}
\value{
Data frame with worker statistics
}
\description{
Get Worker Statistics from Log
}
