% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-error_recovery.R
\name{handle_graceful_degradation}
\alias{handle_graceful_degradation}
\title{Graceful Degradation Handler}
\usage{
handle_graceful_degradation(resource_manager, requested_workers)
}
\arguments{
\item{resource_manager}{Resource manager object}

\item{requested_workers}{Requested number of workers}
}
\value{
Adjusted number of workers
}
\description{
Handles graceful degradation when resources are constrained.
}
\keyword{internal}
