% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariates-similarity.R
\name{kennard_stone_cluster}
\alias{kennard_stone_cluster}
\title{Kennard-Stone Selection for Single Cluster}
\usage{
kennard_stone_cluster(
  cluster_center,
  cluster_unknowns,
  relevant_matrix,
  n_select
)
}
\arguments{
\item{cluster_center}{Numeric vector of cluster centroid in PCA space}

\item{cluster_unknowns}{Matrix of unknown samples in this cluster}

\item{relevant_matrix}{Matrix of relevant OSSL PCA scores}

\item{n_select}{Integer. Number of samples to select}
}
\value{
Vector of indices into relevant_matrix
}
\description{
Core Kennard-Stone algorithm implementation for selecting OSSL samples
most relevant to a specific unknown cluster.
}
\keyword{internal}
