% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-parallel_monitoring.R
\name{log_new_best_model}
\alias{log_new_best_model}
\title{Track Best Model in Log}
\usage{
log_new_best_model(log_file, model_id, rsq, previous_best = NULL)
}
\arguments{
\item{log_file}{Path to log file}

\item{model_id}{Model identifier}

\item{rsq}{R-squared value}

\item{previous_best}{Previous best R-squared}
}
\description{
Logs when a new best model is found
}
