% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation-local-helpers.R
\name{perform_memory_cleanup}
\alias{perform_memory_cleanup}
\title{Perform Memory Cleanup}
\usage{
perform_memory_cleanup(verbose = FALSE)
}
\arguments{
\item{verbose}{Print cleanup messages}
}
\description{
Performs garbage collection and memory defragmentation.
Always uses aggressive cleanup for production quality.
}
\keyword{internal}
