% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-warmstart.R
\name{point_in_grid}
\alias{point_in_grid}
\title{Check if a single-row tibble exists in a grid}
\usage{
point_in_grid(point, grid)
}
\description{
Check if a single-row tibble exists in a grid
}
\keyword{internal}
