% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config-nested_parallel.R
\name{print.nested_parallel_config}
\alias{print.nested_parallel_config}
\title{Print Nested Parallel Configuration}
\usage{
\method{print}{nested_parallel_config}(x, ...)
}
\arguments{
\item{x}{A nested_parallel_config object}

\item{...}{Additional arguments (unused)}
}
\description{
Print Nested Parallel Configuration
}
