% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation-local-helpers.R
\name{print_model_info}
\alias{print_model_info}
\title{Print Model Info}
\usage{
print_model_info(config_row)
}
\arguments{
\item{config_row}{Model configuration row}
}
\description{
Prints formatted information about current model configuration
}
\keyword{internal}
