% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-resource_management.R
\name{release_resource_token}
\alias{release_resource_token}
\title{Release Resource Token}
\usage{
release_resource_token(resource_manager, token)
}
\arguments{
\item{resource_manager}{Resource manager object}

\item{token}{Resource token to release}
}
\description{
Releases a resource token after model evaluation completes.
}
\keyword{internal}
