% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation-nested_parallel_orchestrator.R
\name{set_nested_thread_control}
\alias{set_nested_thread_control}
\title{Set Thread Control for Nested Worker}
\usage{
set_nested_thread_control()
}
\description{
Set Thread Control for Nested Worker
}
\keyword{internal}
