% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-thread_control.R
\name{set_single_thread_mode}
\alias{set_single_thread_mode}
\title{Set Single Thread Mode}
\usage{
set_single_thread_mode(verbose = TRUE)
}
\arguments{
\item{verbose}{Logical. Print status messages (default = TRUE)}
}
\value{
Invisible NULL
}
\description{
Comprehensively sets all threading controls to single-threaded mode.
This prevents thread explosion when running parallel workers.
}
