% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation-local-helpers.R
\name{setup_cv_parallel}
\alias{setup_cv_parallel}
\title{Setup CV Parallel Backend}
\usage{
setup_cv_parallel(n_cores, verbose = TRUE)
}
\arguments{
\item{n_cores}{Number of cores to use}

\item{verbose}{Print setup messages}
}
\description{
Configures parallel backend for cross-validation
}
\keyword{internal}
