% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_join_covariates.R
\name{step_add_covariates}
\alias{step_add_covariates}
\title{Add Scaled Covariates to Recipe Dataset}
\usage{
step_add_covariates(
  recipe,
  covariate_data,
  role = "predictor",
  trained = FALSE,
  skip = FALSE,
  id = recipes::rand_id("add_covariates"),
  sample_id_column = NULL
)
}
\arguments{
\item{recipe}{A \code{recipe} object.}

\item{covariate_data}{A data frame of covariates to be joined.}

\item{role}{Character. Variable role for new covariates (default: "predictor").}

\item{trained}{Logical. Has the step been trained? (required by \code{recipes}).}

\item{skip}{Logical. Should this step be skipped when baking?}

\item{id}{Character. Unique step identifier.}

\item{sample_id_column}{Optional. Name of sample ID column. If NULL, will attempt
to detect it based on variables with role = "id".}
}
\value{
A \code{step_add_covariates} step added to the recipe.
}
\description{
Adds scaled covariates to a recipe dataset via left join using
a shared Sample ID column. Scales covariate data during \code{prep()} and joins
it to the dataset at \code{bake()} time.
}
\examples{
\dontrun{
recipe(data = soil_data) \%>\%
  update_role(Sample_ID, new_role = "id") \%>\%
  step_add_covariates(covariate_data = predicted_covs)
}
}
