% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-parallel_logging.R
\name{tail_parallel_log}
\alias{tail_parallel_log}
\title{Tail Parallel Progress Log}
\usage{
tail_parallel_log(n = 20, log_file = NULL)
}
\arguments{
\item{n}{Integer. Number of lines to show (default: 20)}

\item{log_file}{Character. Path to log file}
}
\description{
Shows the last N lines of the parallel progress log
}
