% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-thread_control.R
\name{thread_control}
\alias{thread_control}
\title{Thread Control Utilities}
\description{
Utilities for managing and verifying thread control in HPC environments.
Ensures single-threaded execution of models to prevent CPU oversubscription.
}
\keyword{control}
\keyword{hpc}
\keyword{thread}
