% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_transform_response.R
\name{transform_outcome}
\alias{transform_outcome}
\title{Transform Outcome Variable Prior to Recipe Construction}
\usage{
transform_outcome(data, method)
}
\arguments{
\item{data}{A data frame containing a column named \code{Response}.}

\item{method}{A character string specifying the transformation to apply. Options include:
\itemize{
\item{"No Transformation"}
\item{"Log Transformation"}
\item{"Square Root Transformation"}
\item{"Box-Cox Transformation"}
}}
}
\value{
A modified data frame with \code{Response} transformed.
}
\description{
Applies a static transformation (e.g., log, sqrt) to the outcome variable (\code{Response})
directly within the input data. This allows compatibility with prediction and workflow
tools that do not handle recipe-based transformations on the outcome.
}
