% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-resource_management.R
\name{update_performance_metrics}
\alias{update_performance_metrics}
\title{Update Performance Metrics}
\usage{
update_performance_metrics(resource_manager, completed, total)
}
\arguments{
\item{resource_manager}{Resource manager object}

\item{completed}{Number of completed models}

\item{total}{Total number of models}
}
\description{
Updates and writes performance metrics to file.
}
\keyword{internal}
