% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-warmstart.R
\name{usable_best_params}
\alias{usable_best_params}
\title{Check whether best_params is usable}
\usage{
usable_best_params(best_params, param_set)
}
\description{
Check whether best_params is usable
}
\keyword{internal}
