% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{VALID_FEATURE_SELECTION}
\alias{VALID_FEATURE_SELECTION}
\title{Valid feature selection methods}
\format{
An object of class \code{character} of length 5.
}
\usage{
VALID_FEATURE_SELECTION
}
\description{
Valid feature selection methods
}
\keyword{internal}
