% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{VALID_PREPROCESSING}
\alias{VALID_PREPROCESSING}
\title{Valid spectral preprocessing methods}
\format{
An object of class \code{character} of length 8.
}
\usage{
VALID_PREPROCESSING
}
\description{
Valid spectral preprocessing methods
}
\keyword{internal}
