% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{VALID_TRANSFORMATIONS}
\alias{VALID_TRANSFORMATIONS}
\title{Valid response transformations}
\format{
An object of class \code{character} of length 3.
}
\usage{
VALID_TRANSFORMATIONS
}
\description{
Valid response transformations
}
\keyword{internal}
