% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-resource_management.R
\name{write_initial_metrics}
\alias{write_initial_metrics}
\title{Write Initial Metrics}
\usage{
write_initial_metrics(resource_manager)
}
\description{
Write Initial Metrics
}
\keyword{internal}
