% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config-hpc.R
\name{calculate_bayes_iter}
\alias{calculate_bayes_iter}
\title{Calculate Recommended Bayesian Iterations}
\usage{
calculate_bayes_iter(inner_workers)
}
\description{
Calculate Recommended Bayesian Iterations
}
\keyword{internal}
