% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation-inner_worker.R
\name{create_workflow_id}
\alias{create_workflow_id}
\title{Create Workflow ID}
\usage{
create_workflow_id(config_row)
}
\arguments{
\item{config_row}{Single row from configuration tibble}
}
\value{
Character string workflow ID
}
\description{
Creates a unique identifier string for a model configuration.
}
\keyword{internal}
