% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onLoad}
\alias{.onLoad}
\title{Package Startup Configuration}
\usage{
.onLoad(libname, pkgname)
}
\description{
Minimal package initialization. Thread control is the user's responsibility.
}
\keyword{internal}
