% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_get_processed_ossl_mir.R
\name{get_processed_mir_path}
\alias{get_processed_mir_path}
\title{Get Path to Processed MIR Spectra}
\usage{
get_processed_mir_path()
}
\value{
Path to the processed MIR spectra \code{.qs} file if it exists, NULL otherwise.
}
\description{
Get Path to Processed MIR Spectra
}
\keyword{internal}
