% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-backtransform.R
\name{needs_back_transformation}
\alias{needs_back_transformation}
\title{Check Whether a Transformation Requires Back-transformation}
\usage{
needs_back_transformation(transformation)
}
\arguments{
\item{transformation}{Character string indicating transformation type.}
}
\value{
Logical.
}
\description{
Check Whether a Transformation Requires Back-transformation
}
\keyword{internal}
