% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_create_project_list.R
\name{project_list}
\alias{project_list}
\title{Define a Named List of Projects for \code{create_input_data()}}
\usage{
project_list(...)
}
\arguments{
\item{...}{Named calls to \code{project_entry()}. Each name will populate the \code{Project} column.}
}
\value{
A named list of project definitions.
}
\description{
Wraps multiple \code{project_entry()} calls into a named list. Each name becomes the
project identifier in the returned dataset.
}
